+++ docs/_pipeline/apps/input-and-gestures/App.cs
@@ -11,8 +11,7 @@ using Windows.System;
 using static Microsoft.UI.Reactor.Factories;
 using static Microsoft.UI.Reactor.Advanced.Factories;
 
-ReactorApp.Run<PointerModifiersExample>("Input and Gestures", width: 640, height: 520
-);
+ReactorApp.Run<PointerModifiersExample>("Input and Gestures");
 
 // <snippet:pointer-modifiers>
 class PointerModifiersExample : Component