+++ docs/_pipeline/apps/recipes-index/App.cs
@@ -2,8 +2,7 @@ using Microsoft.UI.Reactor;
 using Microsoft.UI.Reactor.Core;
 using static Microsoft.UI.Reactor.Factories;
 
-ReactorApp.Run<RecipesIndexApp>("Recipes Gallery", width: 520, height: 320
-);
+ReactorApp.Run<RecipesIndexApp>("Recipes Gallery");
 
 // <snippet:app>
 class RecipesIndexApp : Component