+++ docs/_pipeline/apps/text-and-media/App.cs
@@ -4,8 +4,7 @@ using static Microsoft.UI.Reactor.Factories;
 using static Microsoft.UI.Reactor.Advanced.Factories;
 using Microsoft.UI.Xaml;
 
-ReactorApp.Run<TextAndMediaApp>("Text and Media", width: 720, height: 1200
-);
+ReactorApp.Run<TextAndMediaApp>("Text and Media");
 
 // <snippet:text-variants>
 class TextVariantsDemo : Component